Make ElementWrapper more composeable
